Extracted interfaces into types.ts
